added more feature

+++ src/main/java/ca/mcgill/ecse223/quoridor/controller/Quoridor223Controller.java
@@ -50,7 +50,7 @@ public class Quoridor223Controller {
 		
 		// create Quoridor game and get users
 		Quoridor quoridor = QuoridorApplication.getQuoridor();
-		Game newGame = new Game(GameStatus.Initializing, MoveMode.WallMove, quoridor);
+		Game newGame = new Game(GameStatus.Initializing, false, MoveMode.WallMove, quoridor);
 
 	}
 
@@ -642,8 +642,10 @@ public class Quoridor223Controller {
 		Game curGame = QuoridorApplication.getQuoridor().getCurrentGame();
 		Player curPlayer = getPlayerByName(playerName);
 		if(curPlayer.equals(curGame.getBlackPlayer())) {
+			curGame.setIsFinished(true);
 			curGame.setGameStatus(GameStatus.WhiteWon);
 		}else {
+			curGame.setIsFinished(true);
 			curGame.setGameStatus(GameStatus.BlackWon);
 		}
 	}
@@ -708,16 +710,17 @@ public class Quoridor223Controller {
 	/**
 	 * @author Le-Li Mao
 	 * @throws InvalidOperationException
-	 * @throws GameIsDrawn 
-	 * @throws GameIsFinished 
-	 * @throws GameNotRunningException 
-	 *
 	 */
-	public static void exitReplayMode() throws InvalidOperationException, GameIsDrawn, GameIsFinished, GameNotRunningException {
+	public static void exitReplayMode() throws InvalidOperationException{
 		if (!isReplay())throw new InvalidOperationException("Not in replay mode.");
 		Game curGame = QuoridorApplication.getQuoridor().getCurrentGame();
-		int ind = findPositionIndex();
 		int size = curGame.getPositions().size()-1;
+		if(curGame.getIsFinished()) {
+			curGame.setCurrentPosition(curGame.getPosition(size));
+			throw new InvalidOperationException("Game is finished. Return to the last position.");
+		}
+		curGame.setGameStatus(GameStatus.Running);
+		int ind = findPositionIndex();
 		for(int i = ind; i<size;i++) {
 			Move move = curGame.getMove(ind);
 			move.delete();
@@ -725,16 +728,9 @@ public class Quoridor223Controller {
 			GamePosition pos = curGame.getPosition(ind+1);
 			pos.delete();
 		}
-		System.out.println(curGame.getMoves().size());
 		QuoridorApplication.CreateNewWhitePawnBehavior().startGame();
 		QuoridorApplication.CreateNewBlackPawnBehavior().startGame();
-		curGame.setGameStatus(Game.GameStatus.Running);
-		if(identifyDraw()) {
-			throw new GameIsDrawn("Game Draw!");
-		}
-		if(identifyWin()) {
-			throw new GameIsFinished(curGame.getCurrentPosition().getPlayerToMove().getUser().getName() + " won");
-		}
+		
 	}
 	
 	public static void jumpToFinalPosition() throws InvalidOperationException {
@@ -1386,22 +1382,12 @@ veLegal(newRow, newCol)) throw new InvalidOperationException(String.format("%s: 
 				}
 			} 
 		}
-		//Player current_player = current_game.getCurrentPosition().getPlayerToMove();
-		//PlayerPosition current_position;
-		//PlayerPosition opponent_position;
-		
-		//if(current_player.equals(current_game.getBlackPlayer())) {
-		//	current_position = current_game.getCurrentPosition().getBlackPosition();
-		//	opponent_position = current_game.getCurrentPosition().getWhitePosition();
-		//}
-		//else {
-		//	current_position = current_game.getCurrentPosition().getWhitePosition();
-		//	opponent_position = current_game.getCurrentPosition().getBlackPosition();
-		//}
 		if(identifyDraw()) {
+			curGame.setIsFinished(true);
 			throw new GameIsDrawn("Game Draw!");
 		}
 		if(identifyWin()) {
+			curGame.setIsFinished(true);
 			throw new GameIsFinished(curGame.getCurrentPosition().getPlayerToMove().getUser().getName() + " won. Congratulation!");
 		}
 		SwitchPlayer();
